+++ HISTORY.md
@@ -11,6 +11,7 @@
 * **Notable Changes:**
     * hide empty custom categories in puzzles and when selecting the "hide empty categories" setting
     * shadow (don't inherit / delegate) scripts that use block-instance variables for sprite-clones
+    * custom reporters without an explicit return data type declaration automatically infer one from their definition script when they are edited
     * changed small gear symbol to 6 spikes (from 8)
     * MQTT extension update, thanks, Simon and Xavier!
 * **Notable Fixes:**
@@ -21,6 +22,9 @@
 * **Translation Updates:**
     * German
     * Chinese
+
+### 2026-08-13
+* byob: automatically infer the return data type for custom reporters that don't yet have one by scanning their definition when editing them
 
 ### 2026-08-12
 * lists: fixed clicking on the down-arrow in a list watcher to pop up the index range menu